refactor: MoveDistance 패키지를 position 아래로 수정
