+++ src/main/java/chess/domain/position/MoveDistance.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.position;
 
 public record MoveDistance(int fileMoveAmount, int rankMoveAmount) {
 }
